Replace long long with qint64
We are using it instead of int64_t because is plays nicer with QVariant.
Also add helper to use QJsonValue::toInteger with Qt 6.

+++ torrentfile.h
@@ -26,8 +26,8 @@ namespace libtremotesf {
         int id{};
 
         std::vector<QString> path{};
-        long long size{};
-        long long completedSize{};
+        qint64 size{};
+        qint64 completedSize{};
         Priority priority{};
         bool wanted{};
     };
